Update Combinatorics Generator.py
Removed the show.info message boxes from the mathematical functions, opted instead to include a label in the UI which  updates with each new calculation.

slight bug that needs addressing is after the calculate button is clicked, the check button that is disabled transfers to a 'normal' state.

Which ever check button is disabled needs to remain so until the user makes the change.